RpcServer: expose server fd
Add a few functions that allows to expose the socket()
fd for inet, because socket() and accept() may be called
in different processes.

Also fix a race condition as a side effect. Previously, if
setupSocketServer is called multiple times in different threads,
mServer may be set to an unspecified value because of race. Now,
ensure mLock is held when setting it in setupExternalServer().

Test: binderRpcTest
Test: binderLibTest
Bug: 182914638
Change-Id: I5489516fc92977aca4c98ee2dcdfa8365f8740c9