binderRpcTest: fix hang
There is a hang in this test on device only for unix domain sockets
caused by b/189272263. It seems sending dec refs (which we have many
of now, since each oneway call corresponds to a ref increase), is
filling up the socket. Reducing the stress to 50% until this bug can be
fixed directly.

Bug: N/A
Test: binderRpcTest OnewayStressTest (several hundred times)
Change-Id: I54808707069d8df028404eaeb3b87ec09d499152