libbinder: Parcel always contains header
FIXME: really messy, experiment only, + fixmes
FIXME: there is an issue here with kernel binder, basically when an
empty parcel (usually reply parcel) is constructed, we don't allocate
any memory here, so the object isn't construted. In this case, we either
need to always allocate additional data for these empty replies OR we
need to keep track of when we have the header allocated separately in
Parcel.

This would be in advance of moving other things into the header (have
things like worksource even on manually created interfaces), and it
could allow us to drop the writeInterfaceToken requirement from
libbinder_ndk.

Bug: 188022639
Test: binderRpcTest passes, but a device won't boot because of the above
problem

Change-Id: I22754614ea51aaf5d7ff79e3302341457690d95e