binder: rust: Fix dropped service memleak
When `Binder<T>` objects die, they call the on_destroy for their class.
On creation, the value was put into a Box. On destruction,
`ptr::drop_in_place` was used to drop them. This means that while `Drop`
was called on T, the box was not freed.

Bug: 190517824
Test: Ran with showmap to confirm operations not leaking
Merged-In: Ie9932cc67016eecff4d24ac0c8fcd1bc7df27b32
Change-Id: Ie9932cc67016eecff4d24ac0c8fcd1bc7df27b32