[BugFix] Fix the abnormal roundedcorner in game scenes;
1. when opening a game application, in addition to creating a surfaceview,
it will also create its sublayer background for surfaceview (background for short).
2. generally speaking, the size change of sublayer and the size of roundedCorner
are all changed with the father. However, some games deal with the buffer of
surfaceview due to power consumption, which leads to unity reducing the buffer
size of surfaceview. This buffer and the scale size of the real size are called
bufferScaleTransform.
3. under the full screen gesture, we slide the game up to the latest task.
ComputeBounds has made bufferScaleTransform for background from his father.
However, getRoundedCornersState in the round drawing process does not take
this into account, which leads to some game applications sliding into the
latest task, except for the roundedCorner in the upper left, There are small
black triangles in the other three fillets, that is, the background of
abnormal cutting of fillets.
4. based on the above problems, we submitted the modification to repair
the black background.

Signed-off-by: xinying1 <xinying1@xiaomi.com>
Change-Id: I3d7295088b4c29ab050ae65bd10c9f6dff3d85d5