libbinder: Status - allow null errors
Java Throwable getMessage can (and will!) return null, and when this
happens, we don't get any error information in the native world. So, for
instance, Java ServiceException won't be reported to native code.

This is technically a backwards compatible change, and we'll have to
watch for breakages, but losing service specific exceptions in such a
clearly egregious way is not acceptable. :)

Fixes: 178861468
Test: boot & aidl_integration_test
Change-Id: If70d8fb445078c2989c4288c01055bb7191a012f