binder: Add getCertificate / addTrustedCerticate.
getCertificate returns the self-signed certificate
on the server context.

addTrustedCertificate adds a previously-known certificate
as trusted by this client context.

This change also creates RpcTransportCtxTls as the base class
of the server and client context classes. RpcTransportCtxTls
implements both addTrustedCertificate and getCertificate. This
is so that server and client can verify each other.

Test: binderRpcTest
Bug: 195166979
Change-Id: I0e4fadd8e3391dc39f551e4b80e9411b16b696ab