[binder] Expose public Rust API at crate top level
Previously the public API for the binder Rust crate was nested in a
`public_api` module and re-exported by AIDL interface crates. This
caused confusion, when using multiple AIDL interfaces in the same file,
as there were multiple options to import the binder API.

Moves the public API to the top level of the binder crate and nests the
AIDL implementation API into a submodule.

Test: atest -p --include-subdirs frameworks/native/binder
Bug: 196056781
Change-Id: I811426fa715a324ebb09f5ce04ee539a1c3b9dfd