Testing setDefaultImpl with RPC
Similar set up to what will be in Auto. Trying to find an easy way to
set up the shim without so much manual boilerplating.
Right now, the default impl seems to work when the shim call the proxy
object that it is serving, but NOT when the client calls the same thing.

Test: atest binderDefaultImplTest

Change-Id: Ia4fc4cad6c7bc90f82a768bcf1d17d98cb321a08