libbinder: RPC avoid poll
FIXME: need to get updated stats for latest version of this patch
FIXME: racey test

One of the major costs of RPC binder right now, compared to kernel
binder, is that we need to make two calls (poll + recv) whereas regular
binder can make one (ioctl) in order to read or write a command.

By removing this, we can get more comparable performance:
<TODO>

However, this also introduces the possibility that a shutdown command is
ignored. Also, this isn't currently implemented for the TLS backend.

Bug: 182940634
Test: binderRpcBenchmark (above)
Change-Id: I5cbaf40e5936bdce04b5f158ceac970e8f6ff2fa