sika: cache: Fix up f16 usage to avoid crashes on GPUs that do not support f16
In commit 628feb70dc75 ("Skip fp16 buffers for GPUs that don't
support fp16 textures."), logic was added to avoid crashes on
hardware that doesn't support fp16.

Unfortunately we've seen a few cases where grallocs that don't
handle fp16 still crash with this fix. In some cases the grallocs
can be modified to support fp16, but not always.

The particular issue is that the creation of f16ExternalBuffer
can fail. Right after that object is created, we create
f16ExternalTexture, using the f16ExternalBuffer as an argument
to the constructor.

If the f16ExternalBuffer was not created properly, creating
the f16ExternalTexture will crash.

This patch reworks the logic slightly, so that we check if the
f16ExternalBuffer was created successfully before trying to
create the f16ExternalTexture.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I1ed282d1b6bd7af325b8a1a377ff5d3432424b80