Use std::set in stead of std::vector for mLayersWithQueuedFrames
When something happens on a layer e.g., gpu hang, it observed on some platform that
the layer has trouble in getting its fence signaled. In this case, the underlying
layer comes to its way in handlePageFlip repeatedly at each v-sync, being added to
mLayersWithQueuedFrames again and again. Later, every apearence of this layer in
mLayersWithQueuedFrames will result in a call to latchBuffer, make CPU almost 100%
busy

So simply changing mLayersWithQueuedFrames to be a std::set eliminates this issue

Change-Id: I053787ba777a79528da6af61de8fbea265a5b663
Tracked-On: