Force recursive restorecon if dir newly created
Currently, while assigning new labels to the contents of a directory,
we first check if the label of the parent directory has changed or
not and then proceed to recursively do a restorecon on the entire
directory. This results in a calling restorecon twice for the parent
directory. This works well for the cases when the directory has
already been created. However, the extra call to check if the label has
changed or not can be avoided when we know that a directory has newly
been created. This change forces a recursive restorecon to all parent
dirs if we know they are newly created.

Test: Flashed changes to a sunfish device and tried creating and
switching in to user and checking the labels.
Bug: 191997617

Change-Id: I6b4f1cb1a8794ad5398d2d203557ff4fbfbb9508