atrace: move to rss_stat_throttled
rss_stat is now unthrottled and just going to spam
the trace. rss_stat_throttled is the equivalent of
rss_stat in older kernels.
Note: this will not work on older kernels. I am
not aware of any active use of rss_stat from atrace
(i.e. not going through perfetto) so it's not worth
having a fallback path here for a kernel path that is
going to disappear over time.

Design doc: go/rss-stat-throttled

Bug: 145972256
Change-Id: I0912eca3c15e1ff5f2201c2a66261b7a9b7247a4