Manually calculate stats for app owning external storage
The stats are calculated using quota APIs for all the apps.
This is going wrong for the app owning external storage, as it includes
the external storage files also. The reason for that is the uid for the
external storage files and the app owning the external storage is the
same. Calculating the size manually avoids the problem.

Bug: http://b/195804288
Test: atest StorageHostTest
Test: atest installd_service_test.cpp
Change-Id: I183c6997375b94dc02562bede03e3a15036efb6e