Implement bugreport retrieval logic
This change adds the necessary dumpstate changes to
support the retrieveBugreport API. This includes a new
IDumpstate method, as well as a retrieveBugreport method
in DumpstateService. There is also some refactoring of
the RunStatus handling to reduce code duplication.

This new flow works similarly to the current flow in that
IncidentCompanionService is invoked, with the bugreport files
only being passed to the caller's FileDescriptors if consent
is granted within the timeout window. The files that will be
returned are set by the caller.

Test: atest CtsRootBugreportTestCases
Bug: 196549673
Change-Id: I64d9f525e6b33221fb32c1a536ce8f88bac63210