Allow world searchable trace event dirs for attaching bpf programs
The tracefs default permissions were updated in the kernel to not set
other permission bits by default. [1] In Android, this causes the attaching
bpf programs to tracepoints to fail because the process doing the
attaching doesn't have the required directory file permissions.

Restore the other permission bits, since Android relies on this behavior and
delegates the access control to sepolicy.

[1] https://lore.kernel.org/r/20210818153038.664127804@goodmis.org/T/#m877ba7509de2bffaf0dae63aa9ad8526027a50a9

Bug: 208892266
Test: adb shell dumpsys gpu --gpumem
Test: libtimeinstate_test
Test: libmeminfo_test
Change-Id: Ie93b4952f5cebd80a8a39e83529f26d37ecd016e