Added project id support to internal app directories
Added support to mark the internal app folders with project id and thus calculate storage using project ids on devices whose
user data is wiped. The user data wiped makes sure that all directories are created with project id.

The rationale for doing this change was to provide a proper solution to b/195804288. It will also increase performance and makes the storage quota calculation easier in the long term.

Bug: b/215154615
Test: atest StorageHostTest
Test: atest tests/installd_service_test.cpp
Change-Id: Ib29af7c4c0e80d5b8902899b38de41a807e1cd72