Fix race in waitpid_with_timeout
The a related race was found in copy pasted code in incidentd. dumpstate
has that race, but also another one. Since dumpstate is multithreaded,
it can't rely on using signals at all, because children could be spawned
from multiple threads, and SIGCHLD is directed to the whole process.
Also, if a process blocks a signal, and multiple instances of that
signal are sent (such as multiple SIGCHLDs), they don't queue up. Only
one can be detected.

This all means using signals to manage children in a multithreaded
environemnt doesn't work. Thankfully, the 5.3 kernel FINALLY introduced
pidfd_open, which lets us obtain a file descriptor that refers to a
process. We can then poll on this with a timeout. No need for any signal
handling.

We might want to consider also applying this change to incidentd so
that it's thread-safe, and because it's overall simpler and more robust.

Test: atest dumpstate_smoke_test
Change-Id: I8324a1eb065d31b9adf1a261281c57f31cf84f84