dumpstate: s/vfork/fork/
Quoting vfork's manpage at
https://man7.org/linux/man-pages/man2/vfork.2.html:

```
The vfork() function has the same effect as fork(2), except that the
behavior is undefined if the process created by vfork() either modifies
any data other than a variable of type pid_t used to store the return
value from vfork(), or returns from the function in which vfork() was
called, or calls any other function before successfully calling _exit(2)
or one of the exec(3) family of functions.
```

There are many function calls between this vfork and the child's exec or
_exit. It seems better to simply call fork and sidestep the undefined
behavior here.

Caught by clang's static analyzer:
> frameworks/native/cmds/dumpstate/DumpstateUtil.cpp:321:17: warning:
Call to function 'vfork' is insecure as it can lead to denial of service
situations in the parent process. Replace calls to vfork with calls to
the safer 'posix_spawn' function
[clang-analyzer-security.insecureAPI.vfork]

Bug: 206470603
Test: TreeHugger
Change-Id: I090f623ad4a101176cc4f1fb02fde1ae4ce23b0b